+++ src/examples/java/mt4/MetaApiRpcExample.java
@@ -25,12 +25,12 @@ import cloud.metaapi.sdk.util.JsonMapper;
  */
 public class MetaApiRpcExample {
 
-    private static String token = getEnvOrDefault("NAME", "<put in your token here>");
+    private static String token = getEnvOrDefault("TOKEN", "<put in your token here>");
     private static String login = getEnvOrDefault("LOGIN", "<put in your MT login here>");
     private static String password = getEnvOrDefault("PASSWORD", "<put in your MT password here>");
     private static String serverName = getEnvOrDefault("SERVER", "<put in your MT server name here>");
     private static String brokerSrvFile = getEnvOrDefault("PATH_TO_BROKER_SRV", "/path/to/your/broker.srv");
-  
+    
     public static void main(String[] args) {
         try {
             MetaApi api = new MetaApi(token);
@@ -42,7 +42,12 @@ public class MetaApiRpcExample {
                 .findFirst();
             if (!profile.isPresent()) {
                 System.out.println("Creating account profile");
-                NewProvisioningProfileDto newDto = new NewProvisioningProfileDto() {{ name = serverName; version = 4; }};
+                NewProvisioningProfileDto newDto = new NewProvisioningProfileDto() {{
+                    name = serverName;
+                    version = 4;
+                    brokerTimezone = "EET";
+                    brokerDSTTimezone = "EET";
+                }};
                 profile = Optional.of(api.getProvisioningProfileApi().createProvisioningProfile(newDto).get());
                 profile.get().uploadFile("broker.srv", brokerSrvFile).get();
             }
@@ -70,7 +75,6 @@ public class MetaApiRpcExample {
                     password = mtPassword;
                     server = serverName;
                     provisioningProfileId = provisioningProfile.getId();
-                    timeConverter = "icmarkets";
                     application = "MetaApi";
                     magic = 1000;
                 }}).get());